+++ CHANGELOG.md
@@ -186,6 +186,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - (API) Removed methods `BlockPropertyType.getProcessor()` and `BlockPropertyType.setProcessor()`.
 - (API) Removed network-related code in `Weather`
 - (API) Removed method `BlockType.register()`.
+- (API) Removed method `GenerateFunction.getName()`.
 - (API) Removed field `BlockPlaceEvent.entity` since the entity can be obtained from the interact info.
 - (API) Removed a number of useless build-in permissions which are listed in class `Permissions`, the developer should listen to the corresponded events.